refactor move and rename component
